refactor: complete modular split of CobaltConverter
- main_frame.py: CobaltConverterFrame using ConversionEngine with callbacks
- __init__.py: exports main()
- Languages/ moved into cobalt_converter package
- CobaltConverter.py reduced to 4-line entry point
