+++ cobalt_converter/__init__.py
@@ -1,0 +1,3 @@
+from cobalt_converter.main_frame import main
+
+__all__ = ["main"]
